requirements.txt
Flask
fastapi
uvicorn
pydantic
grpcio
grpcio-tools